email integration added

+++ Movie-ticket-booking/Book-My-Movie/src/main/java/com/bookmymovie/dto/ResponseDtos/TicketResponseDto.java
@@ -21,5 +21,5 @@ public class TicketResponseDto {
     private String theaterName;
     private String address;
     private String bookedSeats;
-    private Integer totalPrice;
+    private double totalPrice;
 }
